+++ test/ids-fieldset/ids-fieldset-e2e-test.ts
@@ -1,4 +1,5 @@
 import { AxePuppeteer } from '@axe-core/puppeteer';
+import countObjects from '../helpers/count-objects';
 
 describe('Ids Fieldset e2e Tests', () => {
   const url = 'http://localhost:4444/ids-fieldset/example.html';
@@ -17,5 +18,16 @@ describe('Ids Fieldset e2e Tests', () => {
 
     const results = await new AxePuppeteer(page).analyze();
     expect(results.violations.length).toBe(0);
+  });
+
+  it('should not have memory leaks', async () => {
+    const numberOfObjects = await countObjects(page);
+    await page.evaluate(() => {
+      document.body.insertAdjacentHTML('beforeend', `<ids-fieldset id="test">
+        <legend><ids-text font-size="20" type="h2" id="user-info">User Information</ids-text></legend>
+      </ids-fieldset>`);
+      document.querySelector('#test')?.remove();
+    });
+    expect(await countObjects(page)).toEqual(numberOfObjects);
   });
 });