+++ test/ids-layout-flex/ids-layout-flex-e2e-test.ts
@@ -1,4 +1,5 @@
 import { AxePuppeteer } from '@axe-core/puppeteer';
+import countObjects from '../helpers/count-objects';
 
 describe('Ids Layout Flex e2e Tests', () => {
   const url = 'http://localhost:4444/ids-layout-flex/example.html';
@@ -16,5 +17,14 @@ describe('Ids Layout Flex e2e Tests', () => {
     await page.goto(url, { waitUntil: ['networkidle2', 'load'] });
     const results = await new AxePuppeteer(page).analyze();
     expect(results.violations.length).toBe(0);
+  });
+
+  it('should not have memory leaks', async () => {
+    const numberOfObjects = await countObjects(page);
+    await page.evaluate(() => {
+      document.body.insertAdjacentHTML('beforeend', ` <ids-layout-flex id="test" gap="8"></ids-layout-flex>`);
+      document.querySelector('#test')?.remove();
+    });
+    expect(await countObjects(page)).toEqual(numberOfObjects);
   });
 });