+++ test/ids-progress-chart/ids-progress-chart-e2e-test.ts
@@ -1,4 +1,5 @@
 import { AxePuppeteer } from '@axe-core/puppeteer';
+import countObjects from '../helpers/count-objects';
 
 describe('Ids Progress Chart e2e Tests', () => {
   const url = 'http://localhost:4444/ids-progress-chart/example.html';
@@ -15,5 +16,15 @@ describe('Ids Progress Chart e2e Tests', () => {
     // both fail axe tests against both light and dark mode backgrounds
     const results = await new AxePuppeteer(page).disableRules(['color-contrast']).analyze();
     expect(results.violations.length).toBe(0);
+  });
+
+  it('should not have memory leaks', async () => {
+    const numberOfObjects = await countObjects(page);
+    await page.evaluate(() => {
+      document.body.insertAdjacentHTML('beforeend', `<ids-progress-chart id="test" label-progress="90%" progress="90" label="dark"></ids-progress-chart>`);
+      document.querySelector('#test')?.remove();
+    });
+
+    expect(await countObjects(page)).toEqual(numberOfObjects);
   });
 });