+++ test/ids-slider/ids-slider-e2e-test.ts
@@ -1,4 +1,5 @@
 import { AxePuppeteer } from '@axe-core/puppeteer';
+import countObjects from '../helpers/count-objects';
 
 describe('Ids Slider e2e Tests', () => {
   const url = 'http://localhost:4444/ids-slider/example.html';
@@ -13,5 +14,15 @@ describe('Ids Slider e2e Tests', () => {
     await page.goto(url, { waitUntil: ['networkidle2', 'load'] });
     const results = await new AxePuppeteer(page).disableRules(['color-contrast']).analyze();
     expect(results.violations.length).toBe(0);
+  });
+
+  it.skip('should not have memory leaks', async () => {
+    const numberOfObjects = await countObjects(page);
+    await page.evaluate(() => {
+      document.body.insertAdjacentHTML('beforeend', `<ids-slider id="test" type="single" show-tooltip label="Single-thumb"></ids-slider>`);
+      document.querySelector('#test')?.remove();
+    });
+
+    expect(await countObjects(page)).toEqual(numberOfObjects);
   });
 });